# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nexora Connect Engine — WhatsApp session management API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  schemas:
...
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      required: [accessToken, refreshToken]
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Exchange a refresh token for a new token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required: [refreshToken]
      responses:
        "200":
          description: New token pair issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/TokenPair"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/logout-all:
    post:
      operationId: logoutAll
      tags: [auth]
      summary: Revoke all refresh tokens for the current user
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        "204":
          description: All sessions revoked
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Current User ──────────────────────────────────────────────────────────────
...